refactor: apply linter's advices
