+++ cdevents-sdk/src/uri_reference.rs
@@ -49,9 +49,9 @@ impl TryFrom<String> for UriReference {
     }
 }
 
-impl ToString for UriReference {
-    fn to_string(&self) -> String {
-        self.0.as_str().to_owned()//into_string()
+impl std::fmt::Display for UriReference {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        self.0.as_str().fmt(f)
     }
 }
